+++ python/python/glide/async_commands/transaction.py
@@ -1551,6 +1551,26 @@ class BaseTransaction:
         """
         return self.append_command(RequestType.DBSize, [])
 
+    def pfadd(self: TTransaction, key: str, elements: List[str]) -> TTransaction:
+        """
+        Adds all elements to the HyperLogLog data structure stored at the specified `key`.
+        Creates a new structure if the `key` does not exist.
+
+        When no `elements` are provided, and `key` exists and is a HyperLogLog, then no operation is performed.
+        If `key` does not exist, then the HyperLogLog structure is created.
+
+        See https://redis.io/commands/pfadd/ for more details.
+
+        Args:
+            key (str): The `key` of the HyperLogLog data structure to add elements into.
+            elements (List[str]): An list of members to add to the HyperLogLog stored at `key`.
+
+        Commands response:
+            int: If the HyperLogLog is newly created, or if the HyperLogLog approximated cardinality is
+            altered, then returns `1`. Otherwise, returns `0`.
+        """
+        return self.append_command(RequestType.PfAdd, [key] + elements)
+
 
 class Transaction(BaseTransaction):
     """